Move DefaultVariableContainerTest to the correct package
